+++ src/modules/commons/index.js
@@ -5,4 +5,4 @@ import './components/checkbox';
 import './components/icon';
 import './components/autocomplete';
 import './components/spinner';
-import './KeyActionMapper';
+
